+++ index.js
@@ -10,7 +10,7 @@ let Characteristic;
 
 const PLUGIN_NAME = 'homebridge-xiaomi-fan';
 const ACCESSORY_NAME = 'xiaomifan';
-const PLUGIN_VERSION = '1.1.4';
+const PLUGIN_VERSION = '1.2.0';
 
 const SMARTMI_MIIO_DEVICES = ['zhimi.fan.sa1', 'zhimi.fan.za1', 'zhimi.fan.za3', 'zhimi.fan.za4','dmaker.fan.p5'];
 const DMAKER_MIIO_DEVICES = ['dmaker.fan.p5'];